+++ shared-mc/src/main/java/dev/blockacademy/tipsign/TipSignModClient.java
@@ -1,0 +1,9 @@
+package dev.blockacademy.tipsign;
+
+import net.fabricmc.api.ClientModInitializer;
+
+public class TipSignModClient implements ClientModInitializer {
+    @Override
+    public void onInitializeClient() {
+    }
+}